ins_type debug

+++ src/libLookup/Lookup.cpp
@@ -4846,7 +4846,8 @@ bool Lookup::ToBlockMessage(unsigned char ins_byte) {
     return true;
   }
   LOG_GENERAL(INFO, "Chetan ToBlockMessage ins_byte="
-                        << ins_byte << " m_syncType=" << m_syncType);
+                        << static_cast<unsigned int>(ins_byte)
+                        << " m_syncType=" << m_syncType);
 
   return m_syncType != SyncType::NO_SYNC &&
          (ins_byte != LookupInstructionType::SETDSBLOCKFROMSEED &&
